requirements.txt
langchain
langgraph
langchain-openai
langchain-community
faiss-cpu
fastapi
uvicorn
python-dotenv
pypdf